+++ mimic-api/src/main/kotlin/MimicService.kt
@@ -30,9 +30,10 @@ public interface MimicService {
         get() = true
 
     /**
-     * Returns the ID of the service.
+     * Returns the ID of the service. It should match regex `[a-z0-9._-]`.
+     *
      * ID should be unique, so it is recommended to use the plugin name as ID.
-     * Default implementation uses plugin name as ID.
+     * Default implementation uses plugin name as an ID.
      */
     public val id: String
         get() = USE_PLUGIN_NAME_AS_ID