+++ mimic-bukkit-api/src/main/kotlin/internal/ItemStack.kt
@@ -1,0 +1,10 @@
+package ru.endlesscode.mimic.internal
+
+import org.bukkit.inventory.ItemStack
+
+internal fun Sequence<ItemStack?>.filterNotNullOrEmpty(): Sequence<ItemStack> {
+    return filterNotNull()
+        .filterNot(ItemStack::isEmpty)
+}
+
+private fun ItemStack.isEmpty(): Boolean = this.type.isAir