+++ mimic-bukkit/src/main/kotlin/WrappedMimicService.kt
@@ -14,16 +14,15 @@ internal abstract class WrappedMimicService(
         get() = pluginManager.isPluginEnabled(pluginName) && delegate.isEnabled
 
     private companion object {
-        val ID_REGEX by lazy { Regex("[a-z0-9_]+") }
+        /**
+         * Matches to Minecraft resource namespace regex.
+         * https://minecraft.fandom.com/wiki/Resource_location#Java_Edition
+         */
+        val ID_REGEX by lazy { Regex("[a-z\\d._-]+") }
 
         fun serviceId(requestedId: String, pluginName: String): String {
             return requestedId.ifEmpty { pluginName }.lowercase().also { id ->
-                require(id matches ID_REGEX) {
-                    """
-                    Invalid ID: $requestedId
-                    It should contain only lowercase Latin letters and digits (a-z, 0-9).
-                    """.trimIndent()
-                }
+                require(id matches ID_REGEX) { "Invalid ID: $requestedId. It should match regex [a-z0-9._-]." }
             }
         }
     }